Conflicts:
	features/events/api/src/main/java/org/opennms/netmgt/events/api/EventParameterUtils.java
	opennms-config/src/main/java/org/opennms/netmgt/config/GroupManager.java
	opennms-config/src/main/java/org/opennms/netmgt/config/UserManager.java
	opennms-services/src/main/java/org/opennms/netmgt/eventd/EventUtil.java
	opennms-services/src/main/java/org/opennms/netmgt/notifd/BroadcastEventProcessor.java
	opennms-services/src/main/java/org/opennms/netmgt/poller/Poller.java
	opennms-services/src/main/java/org/opennms/netmgt/scriptd/helper/SnmpTrapHelper.java
	opennms-services/src/main/java/org/opennms/netmgt/utils/FileBackup.java
	opennms-services/src/main/java/org/opennms/netmgt/xmlrpcd/XmlRpcNotifier.java
	opennms-services/src/main/resources/META-INF/opennms/applicationContext-discovery.xml
	opennms-services/src/test/java/org/opennms/netmgt/poller/PollerTest.java
	opennms-webapp/src/main/java/org/opennms/web/nodelabel/NodeLabelChangeServlet.java
	opennms-webapp/src/main/webapp/admin/sched-outages/editoutage.jsp

+++ opennms-services/src/main/java/org/opennms/netmgt/collectd/SnmpIfData.java
@@ -48,7 +48,7 @@ public class SnmpIfData {
     private final String m_ifAlias;
     
     private final Map<String,String> m_attributes = new HashMap<String,String>();
-
+    
     /**
      * <p>Constructor for SnmpIfData.</p>
      *
@@ -76,8 +76,8 @@ public class SnmpIfData {
         return (num == null ? dflt : num.intValue());
     }
 
-    protected static long nullSafeUnbox(Long num, int dflt) {
-        return (num == null ? dflt : num.intValue());
+    protected static long nullSafeUnbox(Long num, long dflt) {
+        return (num == null ? dflt : num.longValue());
     }
 
     /**
